fix(ffi): pin FFI enum widths explicitly instead of assuming -fshort-enums
The Rust FFI bindgen passed -fshort-enums but no committed C++ build config
does, so every enum-bearing POD crossing include/libdeluge/*.h was laid out
inconsistently between the two sides. Remove the flag (bindgen now agrees
with the C++ build's actual default int-sized enums for everything except
DelugeRegionState, which crosses the FFI by value and gets an explicit
: uint8_t fixed underlying type to match its deliberately-narrow Rust
#[repr(u8)]) — this is what made the icoustic golden diverge (a garbage
4-byte read of a 1-byte C++ return). Update every hand-written host-side
enum stand-in and ABI static assert (control.rs, sd.rs, sys_host.rs) to the
new int-sized reality, apply the same bindgen fix to lens1_vt_sim's build
script, and fix region_differential's own hand-rolled extern declarations
(which would otherwise have silently regressed from correct to wrong the
moment DelugeRegionState went narrow).

+++ src/bsp/rust/region_differential/src/cpp_backend.rs
@@ -45,14 +45,20 @@ extern "C" {
         stream_backing: *mut c_void,
         geometry: DelugeSampleGeometry,
     ) -> *mut DelugeSampleSource;
+    // Return type `u8`, NOT `i32`: `DelugeRegionState` (`include/libdeluge/
+    // sample_source.h`) has an explicit `: uint8_t` fixed underlying type
+    // (deliberately narrow, since it crosses the FFI by value) — an `i32`
+    // declaration here would read 3 bytes of undefined register content above
+    // the true 1-byte return, exactly the width-mismatch bug this whole port
+    // exists to catch. See `RegionState::from_raw` (ops.rs).
     fn deluge_sample_region_acquire_ex(
         src: *mut DelugeSampleSource,
         index: u32,
         direction: i8,
         priority: u32,
         out: *mut DelugeSampleRegion,
-    ) -> i32;
-    fn deluge_sample_region_state(src: *const DelugeSampleSource, index: u32) -> i32;
+    ) -> u8;
+    fn deluge_sample_region_state(src: *const DelugeSampleSource, index: u32) -> u8;
     fn deluge_sample_region_retain(lease: u64);
     fn deluge_sample_region_release(lease: u64);
     fn deluge_sample_source_close(src: *mut DelugeSampleSource);
